Manages to display articles on the page
